requirements.txt
pandas
numpy
transformers
torch
datasets
faiss-cpu
langchain
openai
spacy
sentence-transformers